new architecture (roll back to ubuntu 18.04)

+++ architecture/create-abstract.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ ! "${WEB_FOLDER}" ]]; then
+  echo "ERROR - You must not call this script directly"
+  exit 1
+fi
+
 HOUR=$(date +%H:%M:%S)
 echo "[${HOUR}]===[CLEAN]==="
 sudo rm -rf ./${WEB_FOLDER}/var
@@ -10,12 +15,16 @@ echo "[${HOUR}]===[PROVISION]==="
 ssh root@${ENV_HOST} -p ${ENV_SSH_PORT} $ENV_FOLDER/architecture/scripts/provision.sh "$ENV_TYPE"
 
 HOUR=$(date +%H:%M:%S)
-echo "[${HOUR}]===[PERMISSION]==="
-ssh root@${ENV_HOST} -p ${ENV_SSH_PORT} $ENV_FOLDER/architecture/scripts/permissions.sh
+echo "[${HOUR}]===[CREATE DATABASE]==="
+ssh root@${ENV_HOST} -p ${ENV_SSH_PORT} $ENV_FOLDER/architecture/scripts/createDb.sh "$ENV_TYPE"
 
 HOUR=$(date +%H:%M:%S)
-echo "[${HOUR}]===[CREATE DATABASE]==="
-ssh root@${ENV_HOST} -p ${ENV_SSH_PORT} $ENV_FOLDER/architecture/scripts/createDb.sh
+echo "[${HOUR}]===[PERMISSION]==="
+ssh root@${ENV_HOST} -p ${ENV_SSH_PORT} $ENV_FOLDER/architecture/scripts/permissions.sh "$ENV_TYPE"
+
+HOUR=$(date +%H:%M:%S)
+echo "[${HOUR}]===[TEST]==="
+ssh ${ENV_USER}@${ENV_HOST} -p ${ENV_SSH_PORT} $ENV_FOLDER/architecture/scripts/test.sh "$ENV_TYPE"
 
 HOUR=$(date +%H:%M:%S)
 echo "[${HOUR}]===[INSTALL]==="
